Extract TestPairPriority
to separate file
